chore: remove legacy constants file
