+++ src/constants/items.js
@@ -1,0 +1,2 @@
+export const ITEM_TYPES = ['weapon', 'armor', 'trinket'];
+export const RARITY_ORDER = { rare: 3, uncommon: 2, common: 1 };
