+++ src/constants/phases.js
@@ -1,0 +1,6 @@
+export const PHASES = {
+  MORNING: 'morning',
+  CRAFTING: 'crafting',
+  SHOPPING: 'shopping',
+  END_DAY: 'end_day',
+};
